split cluster into individual files
