+++ internal/cmd/cluster/command.go
@@ -2,34 +2,35 @@ package cluster
 
 import (
 	"github.com/spf13/cobra"
+	"github.com/spf13/pflag"
 
 	pcmd "github.com/confluentinc/cli/internal/pkg/cmd"
+	"github.com/confluentinc/cli/internal/pkg/log"
 )
+
+var subcommandFlags = map[string]*pflag.FlagSet{
+	"list":       pcmd.ContextSet(),
+	"register":   pcmd.ContextSet(),
+	"unregister": pcmd.ContextSet(),
+}
 
 type command struct {
 	*pcmd.StateFlagCommand
-	prerunner  pcmd.PreRunner
-	metaClient Metadata
 }
 
-// New returns the Cobra command for `cluster`.
-func New(prerunner pcmd.PreRunner, metaClient Metadata) *cobra.Command {
-	cmd := &command{
-		StateFlagCommand: pcmd.NewAnonymousStateFlagCommand(&cobra.Command{
-			Use:         "cluster",
-			Short:       "Retrieve metadata about Confluent Platform clusters.",
-			Annotations: map[string]string{pcmd.RunRequirement: pcmd.RequireOnPremLogin},
-		}, prerunner, SubcommandFlags),
-		prerunner:  prerunner,
-		metaClient: metaClient,
+func New(prerunner pcmd.PreRunner, userAgent string, logger *log.Logger) *cobra.Command {
+	cmd := &cobra.Command{
+		Use:         "cluster",
+		Short:       "Retrieve metadata about Confluent Platform clusters.",
+		Annotations: map[string]string{pcmd.RunRequirement: pcmd.RequireOnPremLogin},
 	}
-	cmd.init()
-	return cmd.Command
-}
 
-func (c *command) init() {
-	c.AddCommand(NewDescribeCommand(c.prerunner, c.metaClient))
-	c.AddCommand(NewListCommand(c.prerunner))
-	c.AddCommand(NewRegisterCommand(c.prerunner))
-	c.AddCommand(NewUnregisterCommand(c.prerunner))
+	c := &command{StateFlagCommand: pcmd.NewAnonymousStateFlagCommand(cmd, prerunner, subcommandFlags)}
+
+	c.AddCommand(newDescribeCommand(prerunner, userAgent, logger))
+	c.AddCommand(newListCommand(prerunner))
+	c.AddCommand(newRegisterCommand(prerunner))
+	c.AddCommand(newUnregisterCommand(prerunner))
+
+	return c.Command
 }
