+++ internal/cmd/cluster/command_describe.go
@@ -1,10 +1,7 @@
 package cluster
 
 import (
-	"encoding/json"
 	"fmt"
-	"io/ioutil"
-	"net/http"
 	"sort"
 
 	"github.com/confluentinc/go-printer"
@@ -19,42 +16,6 @@ import (
 	"github.com/confluentinc/cli/internal/pkg/utils"
 )
 
-type Metadata interface {
-	DescribeCluster(url string, caCertPath string) (*ScopedId, error)
-}
-
-type ScopedId struct {
-	ID    string `json:"id"`
-	Scope *Scope `json:"scope"`
-}
-
-type Scope struct {
-	// Path defines the "outer scope" which isn't used yet. The hierarchy
-	// isn't represented in the Scope object in practice today
-	Path []string `json:"path"`
-	// Clusters defines all the key-value pairs needed to uniquely identify a scope
-	Clusters map[string]string `json:"clusters"`
-}
-
-type Element struct {
-	Type string `json:"type" yaml:"type"`
-	ID   string `json:"id" yaml:"id"`
-}
-
-// ScopedIdService allows introspecting details from a Confluent cluster.
-// This is for querying the endpoint each CP service exposes at /v1/metadata/id.
-type ScopedIdService struct {
-	userAgent string
-	logger    *log.Logger
-}
-
-func NewScopedIdService(userAgent string, logger *log.Logger) *ScopedIdService {
-	return &ScopedIdService{
-		userAgent: userAgent,
-		logger:    logger,
-	}
-}
-
 var (
 	describeFields = []string{"Type", "ID"}
 	describeLabels = []string{"Type", "ID"}
@@ -62,67 +23,44 @@ var (
 
 type describeCommand struct {
 	*pcmd.CLICommand
-	client Metadata
+	client metadata
 }
 
-// NewDescribeCommand returns the sub-command object for describing clusters through /v1/metadata/id
-func NewDescribeCommand(prerunner pcmd.PreRunner, client Metadata) *cobra.Command {
-	describeCmd := &describeCommand{
-		CLICommand: pcmd.NewAnonymousCLICommand(&cobra.Command{
-			Use:   "describe",
-			Short: "Describe a Kafka cluster.",
-			Long:  fmt.Sprintf("Describe a Kafka cluster.\nEnvironment variable `%s` can replace the `--url` flag, and `%s` can replace the `--ca-cert-path` flag.", pauth.ConfluentPlatformMDSURL, pauth.ConfluentPlatformCACertPath),
-			Args:  cobra.NoArgs,
-			Example: examples.BuildExampleString(
-				examples.Example{
-					Text: "Discover the cluster ID and Kafka ID for Connect.",
-					Code: "confluent cluster describe --url http://localhost:8083",
-				},
-			),
-		}, prerunner),
-		client: client,
-	}
-	describeCmd.Flags().String("url", "", "URL to a Confluent cluster.")
-	describeCmd.Flags().String("ca-cert-path", "", "Self-signed certificate chain in PEM format.")
-	describeCmd.Flags().StringP(output.FlagName, output.ShortHandFlag, output.DefaultValue, output.Usage)
-	describeCmd.RunE = pcmd.NewCLIRunE(describeCmd.describe)
-
-	return describeCmd.Command
+type metadata interface {
+	DescribeCluster(url, caCertPath string) (*ScopedId, error)
 }
 
-func (s *ScopedIdService) DescribeCluster(url string, caCertPath string) (*ScopedId, error) {
-	var httpClient *http.Client
-	if caCertPath != "" {
-		var err error
-		httpClient, err = utils.SelfSignedCertClientFromPath(caCertPath, s.logger)
-		if err != nil {
-			return nil, err
-		}
-	} else {
-		httpClient = utils.DefaultClient()
+type element struct {
+	Type string `json:"type" yaml:"type"`
+	ID   string `json:"id" yaml:"id"`
+}
+
+func newDescribeCommand(prerunner pcmd.PreRunner, userAgent string, logger *log.Logger) *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "describe",
+		Short: "Describe a Kafka cluster.",
+		Long:  fmt.Sprintf("Describe a Kafka cluster. Environment variable `%s` can replace the `--url` flag, and `%s` can replace the `--ca-cert-path` flag.", pauth.ConfluentPlatformMDSURL, pauth.ConfluentPlatformCACertPath),
+		Args:  cobra.NoArgs,
+		Example: examples.BuildExampleString(
+			examples.Example{
+				Text: "Discover the cluster ID and Kafka ID for Connect.",
+				Code: "confluent cluster describe --url http://localhost:8083",
+			},
+		),
 	}
-	ctx := utils.GetContext(s.logger)
-	req, err := http.NewRequestWithContext(ctx, "GET", fmt.Sprintf("%s/v1/metadata/id", url), nil)
-	if err != nil {
-		return nil, err
+
+	c := &describeCommand{
+		CLICommand: pcmd.NewAnonymousCLICommand(cmd, prerunner),
+		client:     newScopedIdService(userAgent, logger),
 	}
-	req.Header.Set("User-Agent", s.userAgent)
-	req.Header.Set("Accept", "application/json")
-	resp, err := httpClient.Do(req)
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-	body, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		return nil, err
-	}
-	if resp.StatusCode != http.StatusOK {
-		return nil, errors.Errorf(errors.FetchClusterMetadataErrorMsg, resp.Status, body)
-	}
-	meta := &ScopedId{}
-	err = json.Unmarshal(body, meta)
-	return meta, err
+
+	c.RunE = pcmd.NewCLIRunE(c.describe)
+
+	c.Flags().String("url", "", "URL to a Confluent cluster.")
+	c.Flags().String("ca-cert-path", "", "Self-signed certificate chain in PEM format.")
+	c.Flags().StringP(output.FlagName, output.ShortHandFlag, output.DefaultValue, output.Usage)
+
+	return c.Command
 }
 
 func (c *describeCommand) describe(cmd *cobra.Command, _ []string) error {
@@ -172,11 +110,11 @@ func getCACertPath(cmd *cobra.Command) (string, error) {
 }
 
 func printDescribe(cmd *cobra.Command, meta *ScopedId, format string) error {
-	type StructuredDisplay struct {
+	structuredDisplay := &struct {
 		Crn   string    `json:"crn" yaml:"crn"`
-		Scope []Element `json:"scope" yaml:"scope"`
-	}
-	structuredDisplay := &StructuredDisplay{}
+		Scope []element `json:"scope" yaml:"scope"`
+	}{}
+
 	if meta.ID != "" {
 		if format == output.Human.String() {
 			utils.Printf(cmd, "Confluent Resource Name: %s\n\n", meta.ID)
@@ -184,22 +122,24 @@ func printDescribe(cmd *cobra.Command, meta *ScopedId, format string) error {
 			structuredDisplay.Crn = meta.ID
 		}
 	}
+
 	var types []string
 	for name := range meta.Scope.Clusters {
 		types = append(types, name)
 	}
 	sort.Strings(types) // since we don't have hierarchy info, just display in alphabetical order
+
 	var data [][]string
 	for _, name := range types {
 		id := meta.Scope.Clusters[name]
-		element := Element{Type: name, ID: id}
+		element := element{Type: name, ID: id}
 		if format == output.Human.String() {
 			data = append(data, printer.ToRow(&element, describeFields))
 		} else {
 			structuredDisplay.Scope = append(structuredDisplay.Scope, element)
 		}
-
 	}
+
 	if format == output.Human.String() {
 		utils.Println(cmd, "Scope:")
 		printer.RenderCollectionTable(data, describeLabels)
@@ -207,10 +147,4 @@ func printDescribe(cmd *cobra.Command, meta *ScopedId, format string) error {
 		return output.StructuredOutput(format, structuredDisplay)
 	}
 	return nil
-}
-
-func check(err error) {
-	if err != nil {
-		panic(err)
-	}
 }
