+++ internal/cmd/cluster/command_list.go
@@ -15,34 +15,33 @@ type listCommand struct {
 	*pcmd.AuthenticatedCLICommand
 }
 
-// NewListCommand returns the sub-command object for listing clusters
-func NewListCommand(prerunner pcmd.PreRunner) *cobra.Command {
-	listCmd := &listCommand{
-		AuthenticatedCLICommand: pcmd.NewAuthenticatedWithMDSCLICommand(
-			&cobra.Command{
-				Use:   "list",
-				Short: "List registered clusters.",
-				Long:  "List clusters that are registered with the MDS cluster registry.",
-			},
-			prerunner),
+func newListCommand(prerunner pcmd.PreRunner) *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "list",
+		Short: "List registered clusters.",
+		Long:  "List clusters that are registered with the MDS cluster registry.",
 	}
-	listCmd.Flags().StringP(output.FlagName, output.ShortHandFlag, output.DefaultValue, output.Usage)
-	listCmd.RunE = pcmd.NewCLIRunE(listCmd.list)
-	return listCmd.Command
-}
 
-func (c *listCommand) createContext() context.Context {
-	return context.WithValue(context.Background(), mds.ContextAccessToken, c.State.AuthToken)
+	c := &listCommand{AuthenticatedCLICommand: pcmd.NewAuthenticatedWithMDSCLICommand(cmd, prerunner)}
+
+	c.RunE = pcmd.NewCLIRunE(c.list)
+
+	c.Flags().StringP(output.FlagName, output.ShortHandFlag, output.DefaultValue, output.Usage)
+
+	return c.Command
 }
 
 func (c *listCommand) list(cmd *cobra.Command, _ []string) error {
-	clusterInfos, response, err := c.MDSClient.ClusterRegistryApi.ClusterRegistryList(c.createContext(), &mds.ClusterRegistryListOpts{})
+	ctx := context.WithValue(context.Background(), mds.ContextAccessToken, c.State.AuthToken)
+	clusterInfos, response, err := c.MDSClient.ClusterRegistryApi.ClusterRegistryList(ctx, &mds.ClusterRegistryListOpts{})
 	if err != nil {
 		return print.HandleClusterError(err, response)
 	}
+
 	format, err := cmd.Flags().GetString(output.FlagName)
 	if err != nil {
 		return err
 	}
+
 	return print.PrintCluster(clusterInfos, format)
 }
